% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-qr.R
\name{geom_qr}
\alias{geom_qr}
\title{Draw QR Codes}
\usage{
geom_qr(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  ...,
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE
)
}
\arguments{
\item{mapping, data, stat, position, ..., na.rm, show.legend, inherit.aes}{see
documentation for \code{ggplot2::geom_point()}}
}
\description{
Draw QR Codes
}
\section{Aesthetics}{


\code{geom_qr()} understands the following aesthetics:
\itemize{
\item{x}
\item{y}
\item{label}
\item{colour}
\item{fill}
\item{size (in inches)}
}
}

