% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom-qr.R
\name{qrGrob}
\alias{qrGrob}
\title{Render text as a QR code raster object}
\usage{
qrGrob(
  label,
  x = unit(0.5, "npc"),
  y = unit(0.5, "npc"),
  size = unit(72, "pt"),
  just = "centre",
  hjust = NULL,
  vjust = NULL,
  interpolate = FALSE,
  default.units = "npc",
  name = NULL,
  gp = gpar(),
  vp = NULL,
  fill = NULL,
  col = NULL
)
}
\arguments{
\item{label}{text to encode into QR Code}

\item{x, y, just, hjust, vjust, interpolate, default.units, name, gp, vp}{See \code{grid::rasterGrob()} documentation}

\item{size}{used for width and height of QR code. To ensure that the QR code
is actually square, \code{size} is usually best to
be in absolute units like 'pt', 'mm', etc.}

\item{fill, col}{fill(bg) and colour(fg). Will override \code{gp} values
if set}
}
\description{
Render text as a QR code raster object
}
